- updated domain model with Lombok annotations;
